move fight extras to fightutil
