+++ src/main/java/Commands/Fight/Util/FightStatTypes.java
@@ -1,4 +1,4 @@
-package Core.Fight;
+package Commands.Fight.Util;
 
 public enum FightStatTypes {
     HP,
