+++ src/main/java/Commands/Fight/Util/FightUserManager.java
@@ -1,4 +1,4 @@
-package Core.Fight;
+package Commands.Fight.Util;
 
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.User;
